CRUD usuarios e CRUD de tipos de consultas com validações dos campos.

+++ src/main/java/com/Squad03/demo/services/UserService.java
@@ -21,8 +21,28 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> getUserById(UUID id) {
-        return userRepository.findById(id);
+    public User getUserById(UUID id) throws Exception {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            return user.get();
+        } else {
+            throw new Exception("Usuário não encontrado com ID: " + id);
+        }
+    }
+
+
+    public User updateUserById(UUID id, User newUserData) throws Exception {
+        Optional<User> existingUser = userRepository.findById(id);
+
+        if (existingUser.isPresent()) {
+            User userToUpdate = existingUser.get();
+            userToUpdate.setName(newUserData.getName());
+            userToUpdate.setEmail(newUserData.getEmail());
+
+            return userRepository.save(userToUpdate);
+        } else {
+            throw new Exception("Usuário não encontrado com ID: " + id);
+        }
     }
 
 
@@ -30,8 +50,12 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(UUID id) {
-        userRepository.deleteById(id);
+    public void deleteUserById(UUID id) throws Exception {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new Exception("Usuário não encontrado com ID: " + id);
+        }
     }
 
 }
